refactor(test): extract shared helpers into test/helpers/
Pulls server lifecycle (spawnServer, waitForServer, waitForExit,
connectClient), fixture builders (createTestSyncFixture,
createScrivener*), and db setup (insertTestScene,
setupReviewBundleTestDb) into test/helpers/{server,fixtures,db}.js.

No logic changes — prerequisite for Phase B test file split.
314/314 tests pass.
