+++ test/helpers/server.js
@@ -1,0 +1,60 @@
+import { spawn } from "node:child_process";
+import { fileURLToPath } from "node:url";
+import path from "node:path";
+import { Client } from "@modelcontextprotocol/sdk/client/index.js";
+import { SSEClientTransport } from "@modelcontextprotocol/sdk/client/sse.js";
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+const ROOT = path.resolve(__dirname, "../..");
+
+export function spawnServer(port, syncDir, extraEnv = {}) {
+  const proc = spawn(
+    process.execPath,
+    ["--experimental-sqlite", path.join(ROOT, "index.js")],
+    {
+      env: {
+        ...process.env,
+        WRITING_SYNC_DIR: syncDir,
+        DB_PATH: ":memory:",
+        HTTP_PORT: String(port),
+        ...extraEnv,
+      },
+      stdio: ["ignore", "ignore", "pipe"],
+    }
+  );
+  proc.on("error", (err) => {
+    throw new Error(`Failed to start server: ${err.message}`);
+  });
+  return proc;
+}
+
+export async function waitForServer(url, retries = 20, delayMs = 200) {
+  for (let i = 0; i < retries; i++) {
+    try {
+      const res = await fetch(`${url}/healthz`);
+      if (res.ok) return;
+    } catch {}
+    await new Promise((r) => setTimeout(r, delayMs));
+  }
+  throw new Error(`Server at ${url} did not become ready`);
+}
+
+export async function waitForExit(proc, timeoutMs = 5000) {
+  return await new Promise((resolve, reject) => {
+    const timeout = setTimeout(
+      () => reject(new Error("Process did not exit in time")),
+      timeoutMs
+    );
+    proc.once("exit", (code, signal) => {
+      clearTimeout(timeout);
+      resolve({ code, signal });
+    });
+  });
+}
+
+export async function connectClient(url) {
+  const c = new Client({ name: "integration-test-client", version: "1.0.0" });
+  const transport = new SSEClientTransport(new URL(`${url}/sse`));
+  await c.connect(transport);
+  return c;
+}
